+++ validation/validation.js
@@ -1,15 +1,20 @@
-// TODO: Alerts in this file: 7
+// TODO: Alerts in this file: ? if multiline comments, 1 if standalone_items, 5 else
 let a = "todo";
 //TODO: Alert
 let b = "todo";
 // TODO Alert
 let c = "todo";
-/* TODO: Alert */
+/* TODO: Alert */ 
 let d = "todo";
-/* TODO: Alert 
-let alertThis = "todo";
+/* TODO: Alert
+let alertThis = "TODO";
 */
 let e = "todo";
 /*
 TODO: Alert */
 let f = "todo";
+let g = "todo"; // TODO: Alert
+
+
+
+/* TODO: Alert */